It was completly broken, now fixed. Please read desc:
dos2unix AOS.sh --this fixed everything, the ending char of the filetypes always saved with an unknown char at the end even if there wasn't a space. My guess is that it was in the conversion from putty to actually saving. Not sure exactly but the dos2unix command after saving it each time solved the issue.